Updated example to work with latest .net 5 template project.
Original example did not run for me, produced a cryptic VS protocol error. Ended up creating a new project and moving the code across for a fresh start.

Also the FetchData in the original example seemed to purely rely on the animateList css class, with the Transition component not actually doing any animation. This update now uses the ShowIntro parameter to fade-in the table as it's being animated on page load.